move commit: Factor out the disconnection params
We can reuse the logic behind caculating the disconnect params.
