+++ crates/but-workspace/src/lib.rs
@@ -44,6 +44,7 @@ pub use tree_manipulation::discard_worktree_changes::discard_workspace_changes;
 pub mod branch;
 
 mod changeset;
+mod graph_manipulation;
 
 /// Utility types for the [`WorkspaceCommit`].
 pub mod commit;
